backend/import/requirements.txt
requests==2.9.1
behave==1.2.5
pyhamcrest>=1.9.0
clint
gzip